src/main/java/com/analyzer/approachHelper/service/ReviewServiceImpl.java
package com.analyzer.approachHelper.service;

import com.analyzer.approachHelper.domain.Review;
import com.analyzer.approachHelper.repository.ReviewRepository;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class ReviewServiceImpl implements ReviewService {

    private ReviewRepository reviewRepository;

    public ReviewServiceImpl(ReviewRepository reviewRepository) {
        this.reviewRepository = reviewRepository;
    }

    @Override
    public List<Review> getReviewsByProductId(String productId) {
        return reviewRepository.findByProductId(productId);
    }
}
